Makes undotree work with plugins that fold diffs.
Some plugins fold diffs by default.  If such a plugin is installed
along with undotree, `diffpanel.Show()` sets the filetype for
diffpanel to `diff`, which results in a big, happy, closed fold.  Then
`undotree.Draw()` helpfully tries to deletes the last line in the diff
buffer, believing it empty.  Except, said line is now the fold.

The patch addresses this problem by opening the fold, if any.  Perhaps
a better fix would be to make sure the deleted line is actually empty
before removing it.