Add canonical include guard.
I :runtime the plugin early in my .vimrc to check whether it is supported, and then the plugin (and especially the :autocmds) gets defined again when Vim sources the regular plugin/undotree.vim.
Add a canonical include guard, and allow to check whether the plugin is actually enabled by using a different guard value.