fix: make undoing/redoing changes in folds visible
Without Undotree, Vim's default behavior upon undoing or redoing a
change in the current buffer is to open any folds that contain that
change so that the change is visible. This behavior is controlled by the
`undo` option in `&fdo`, which is enabled by default. Undotree is
currently keeping such folds closed when undoing and redoing changes
from the Undotree panel, which ultimately hides the changes that are
being undone or redone.

This commit:

* Brings Undotree's behavior in line with how undo/redo usually works
  (e.g. when using `u` or `CTRL-R` in the current buffer)

* Let's you view undo/redo changes in real time and without having to
  open or disable folds first, even if those changes that you are
  undoing/redoing (from the Undotree panel) are currently hidden in
  closed folds

* Respects Vim's `fdo` option, so that the user may customize this
  behavior